commands: move new code to new.rs
